requirements.txt
alembic==1.7.4
attrs==21.2.0
bcrypt==3.2.0
category-encoders==2.2.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==35.0.0
distro==1.6.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
paramiko==2.7.2
patsy==0.5.2
psycopg2-binary==2.9.1
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.25
statsmodels==0.13.0
texttable==1.6.4
threadpoolctl==3.0.0
urllib3==1.26.7
websocket-client==0.59.0
Werkzeug==2.0.2
wtf==0.1
WTForms==2.3.3
zipp==3.6.0
